Bei 2-OPT weitergearbeitet. Swap funktion und Algorithmus fertig programmiert (sollte alles funktionieren) Einzigstes Problem ist, dass die Randomzahl oft nicht Random ist, weil das Programm zu schnell ist.
